Added key implementation of transactions
Integrate transactions that were accumulated in list of transactions just before they are added to the block.
Make a request to add a new block to the blockchain with a POST request within Postman. Do this making a json file to post the json format into the response, tells us which block was added as well

** Added VARIABLE: node_address **
- for whenever a miner mines a block they get transaction fees from that block
- remembering the add_node method. uuid4() method will generate a unique address... NOTICE UUID function are generated as: {12345678-1234-5678-1234-567812345678} so we will want to replace the '-'s with nothing

** Added KEY: transactions to mining block section **